UPDATE: 将 dataSize 和 numericPrecision 类型修正为 Int FIX: 修正数字的 min max 范围，并采用字符串传递以避免精度问题和科学计数法展示问题 UPDATE: 移除 EditFormRule 中的 id 属性

+++ src/main/kotlin/top/potmot/core/database/generate/columnType/ColumnTypeDefiner.kt
@@ -18,9 +18,9 @@ interface ColumnTypeDefiner {
 
     fun requiredNumericPrecision(typeCode: Int): Boolean
 
-    fun defaultDataSize(typeCode: Int): Long?
+    fun defaultDataSize(typeCode: Int): Int?
 
-    fun defaultNumericPrecision(typeCode: Int): Long?
+    fun defaultNumericPrecision(typeCode: Int): Int?
 
     @Throws(ColumnTypeException::class)
     fun getTypeDefine(typeMeta: ColumnTypeMeta): String =
@@ -35,11 +35,11 @@ interface ColumnTypeDefiner {
     @Throws(ColumnTypeException.MissRequiredParam::class)
     fun getTypeDataSizeAndNumericPrecision(
         typeCode: Int,
-        dataSize: Long? = null,
-        numericPrecision: Long? = null
+        dataSize: Int? = null,
+        numericPrecision: Int? = null
     ): String =
         buildString {
-            var tempDataSize: Long? = null
+            var tempDataSize: Int? = null
 
             val isNeedDataSize = needDataSize(typeCode)
             val isRequiredDataSize = requiredDataSize(typeCode)
@@ -56,7 +56,7 @@ interface ColumnTypeDefiner {
                 }
             }
 
-            var tempNumericPrecision: Long? = null
+            var tempNumericPrecision: Int? = null
 
             val isNeedNumericPrecision = needNumericPrecision(typeCode)
             val isRequiredNumericPrecision = requiredNumericPrecision(typeCode)
